+++ lib/kitchen/directions/bake_numbered_table/bake_table_body.rb
@@ -73,6 +73,9 @@ module Kitchen
           elsif table.unstyled?
             custom_table = CustomBody.new(table: table, klass: 'unstyled')
             custom_table.modify_body(has_fake_title: false)
+          elsif table.data_table?
+            custom_table = CustomBody.new(table: table, klass: 'data-table')
+            custom_table.modify_body(has_fake_title: false)
           end
         end
       end